Added simple request processing
Added:
- Post method for adding car object in database

+++ CarRentalSystem/CarRentalSystem.Infrastructure.Data/CarRentalSystemContext.cs
@@ -8,6 +8,8 @@ namespace CarRentalSystem.Infrastructure.Data
 {
     public sealed class CarRentalSystemContext: DbContext
     {
+        private static bool _isContextHaveBeenCreated;
+
         public DbSet<PointOfRental> PointOfRentals { get; set; }
 
         public DbSet<Car> Cars { get; set; }
@@ -23,8 +25,12 @@ namespace CarRentalSystem.Infrastructure.Data
         public CarRentalSystemContext(DbContextOptions<CarRentalSystemContext> options)
             :base(options)
         {
-            Database.EnsureDeleted();
-            Database.EnsureCreated();
+            if (!_isContextHaveBeenCreated)
+            {
+                Database.EnsureDeleted();
+                Database.EnsureCreated();
+                _isContextHaveBeenCreated = true;
+            }
         }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
